changed dir layout, config bug fixed

+++ server/src/main/java/bernardic/jb/server/Configs.java
@@ -1,4 +1,4 @@
-package com.janbernardic.server;
+package main.java.bernardic.jb.server;
 
 import java.io.FileInputStream;
 import java.io.IOException;
@@ -22,30 +22,19 @@ public class Configs {
 		saveConfig("database");
 	}
 	
-	@SuppressWarnings("resource")
 	public static void saveConfig(String name) {
 		try {
 			Path path = Paths.get("./"+name+".properties");
-			InputStream is;
-			try {
-				is = new FileInputStream(path.toString());
-			} 
-			catch (IOException e1) {
-				is = Configs.class.getResourceAsStream("/resources/configs/" + path.getFileName());
-				try {
-					Files.copy(is, path);
-				} catch (IOException e) {
-					e.printStackTrace();
-				}
+			if(!Files.exists(path)) {
+				InputStream is = Configs.class.getResourceAsStream("/main/resources/configs/" + path.getFileName());
+				Files.copy(is, path);
+				is.close();
 			}
+			InputStream is = new FileInputStream(path.toString());
 			Properties props = new Properties();
 			props.load(is);
 			is.close();
 			configs.put(name, props);
-		}catch(IOException e) {}
-		
-
-
-		
+		}catch(IOException e) {}	
 	}
 }
